### 2025-03-19 77bd0cb
first version of least squares
