+++ src/debiased_spatial_whittle/old_models.py
@@ -1,0 +1,1410 @@
+try:
+    import torch
+except ModuleNotFoundError:
+    pass
+
+from debiased_spatial_whittle.backend import BackendManager
+
+np = BackendManager.get_backend()
+fftn, ifftn = BackendManager.get_fft_methods()
+arange = BackendManager.get_arange()
+
+import numpy
+
+# In this file we define covariance models
+from abc import ABC, abstractmethod
+
+from typing import Tuple, List, Dict, Union
+
+
+class Parameter:
+    """
+    Class designed to handle parameters of model. Note that a Parameter can have it own core value, or might just
+    be a pointer to another parameter. In particular, two parameters might point to a single other parameter.
+    In that case, those two parameters always have the same value.
+    """
+
+    def __init__(self, name: str, bounds: Tuple[float, float]):
+        # TODO hide point_to
+        # TODO add registry to ensure no duplicate names? Add access to registry by name of parameter?
+        self.point_to = None
+        self.name = name
+        self.bounds = bounds
+        self.value = None
+        self.init_guess = 0.9
+        # fixme log scale not expected to work with gradients right now
+        # might need a whole new class LogParameter
+        self.log_scale = False
+
+    # TODO add property name and make it point to if adequate
+
+    @property
+    def value(self):
+        if self.point_to is not None:
+            return self.point_to.value
+        if self._value is None:
+            return None
+        if self.log_scale:
+            return numpy.exp(self._value)
+        return self._value
+
+    @value.setter
+    def value(self, v: float):
+        if self.point_to is not None:
+            self.point_to.value = v
+        self._value = v
+
+    @property
+    def free(self):
+        return self.value is None
+
+    def merge_with(self, other):
+        assert (
+            self.free and other.free
+        ), "Only free parameters (value set to None) can be merged together."
+        new_name = self.name + " and " + other.name
+        new_min = max(self.bounds[0], other.bounds[0])
+        new_max = min(self.bounds[1], other.bounds[1])
+        new_bounds = (new_min, new_max)
+        new_parameter = Parameter(new_name, new_bounds)
+        self.point_to = new_parameter
+        other.point_to = new_parameter
+        return new_parameter
+
+    def __pow__(self, power, modulo=None):
+        return self.value**power
+
+    def __mul__(self, other):
+        return self.value * other
+
+    def __rmul__(self, other):
+        return self.value * other
+
+    def __invert__(self):
+        return 1 / self.value
+
+    def __truediv__(self, other):
+        return self.value / other
+
+    def __rtruediv__(self, other):
+        return other / self.value
+
+    def __repr__(self):
+        return f"{self.name}: {self.value} ... {self.bounds}"
+
+
+class Parameters:
+    """Wrapper for a dictionary of parameters. The keys are the names of the parameters in the model, but two
+    different keys could point to a single Parameter object"""
+
+    def __init__(self, parameters: List[Parameter]):
+        # TODO make param_dict hidden
+        self.param_dict = dict([(p.name, p) for p in parameters])
+
+    @property
+    def names(self):
+        return list(self.param_dict.keys())
+
+    @property
+    def values(self):
+        return [p.value for p in self]
+
+    @property
+    def bounds(self):
+        return [p.bounds for p in self]
+
+    @property
+    def init_guesses(self):
+        return [p.init_guess for p in self]
+
+    def free_params(self):
+        list_free = list(filter(lambda p: p.free, self))
+        return Parameters(list_free)
+
+    def __getitem__(self, item):
+        return self.param_dict[item]
+
+    # TODO add some checks that no two parameters have the same name
+
[…]
