### 2025-03-19 6553967
fix nugget model, fix least square for cupy
