+++ docs/least_squares.py
@@ -1,7 +1,6 @@
 from debiased_spatial_whittle.backend import BackendManager
-from debiased_spatial_whittle.new_models import NuggetModel
 
-BackendManager.set_backend("numpy")
+BackendManager.set_backend("cupy")
 
 np = BackendManager.get_backend()
 
@@ -12,8 +11,9 @@ from debiased_spatial_whittle.periodogram import Periodogram, ExpectedPeriodogra
 from debiased_spatial_whittle.least_squares import LeastSquareEstimator
 from debiased_spatial_whittle.likelihood import DebiasedWhittle, Estimator
 
-grid = RectangularGrid((256, 256))
+grid = RectangularGrid((512, 512))
 model = SquaredExponentialModel(rho=12.0, sigma=2.0)
+model = NuggetModel(model, nugget=1e-3)
 sampler = SamplerOnRectangularGrid(model, grid)
 data = sampler()
 
@@ -21,21 +21,19 @@ periodogram = Periodogram()
 expected_periodogram = ExpectedPeriodogram(grid, periodogram)
 
 model_est = SquaredExponentialModel(rho=10.0)
-model_est_ = NuggetModel(model_est, nugget=1e-5)
-model_est_.fix_parameter("nugget")
-model_est_.fix_parameter("sigma")
+model_est_ = NuggetModel(model_est, nugget=1e-1)
 
 debiased_whittle = DebiasedWhittle(periodogram, expected_periodogram)
 
 least_square = LeastSquareEstimator(periodogram, expected_periodogram)
 least_square(data, model_est_)
-print(model_est.rho, model_est.sigma)
+print(model_est.rho, model_est.sigma, model_est_.nugget)
 print(debiased_whittle(data, model_est_))
 
 
 estimator = Estimator(debiased_whittle)
 estimator(model_est_, data)
-print(model_est.rho, model_est.sigma)
+print(model_est.rho, model_est.sigma, model_est_.nugget)
 print(debiased_whittle(data, model_est_))
 
 model_est.sigma = 1
@@ -45,18 +43,3 @@ print(np.sqrt(np.mean(p / ep)))
 
 model_est.sigma = np.sqrt(np.mean(p / ep))
 print(debiased_whittle(data, model_est_))
-
-print(
-    model_est_(
-        np.array(
-            [
-                [
-                    0,
-                ],
-                [
-                    0,
-                ],
-            ]
-        )
-    )
-)
