### 2025-03-19 3bf354c
add anisotropic model
