### 2025-03-20 59d03c0
add matern32, matern52 and rational quadratic covariance kernels
