+++ src/debiased_spatial_whittle/models.py
@@ -401,6 +401,18 @@ class ExponentialModel(CovarianceModel):
 
 
 class SquaredExponentialModel(CovarianceModel):
+    """
+    Implements the Squared Exponential covariance model, or Gaussian covariance model.
+
+    Attributes
+    ----------
+    rho: float
+        length scale parameter
+
+    sigma: float
+        amplitude parameter
+    """
+
     rho = ModelParameter(default=1.0, bounds=(0, numpy.infty), doc="Range parameter")
     sigma = ModelParameter(
         default=1.0, bounds=(0, numpy.infty), doc="Amplitude parameter"
@@ -432,9 +444,99 @@ class SquaredExponentialModel(CovarianceModel):
         return dict(rho=d_rho, sigma=d_sigma)
 
 
+class Matern32Model(CovarianceModel):
+    """
+    Implements the Matern Covariance kernel with slope parameter 3/2.
+
+    Attributes
+    ----------
+    rho: float
+        length scale parameter of the kernel
+
+    sigma: float
+        amplitude parameter of the kernel
+    """
+
+    rho = ModelParameter(default=1.0, bounds=(0, np.infty))
+    sigma = ModelParameter(default=1.0, bounds=(0, np.infty))
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def _compute(self, lags: np.ndarray):
+        d = np.sqrt(np.sum(lags**2, 0))
+        return (
+            self.sigma**2
+            * (1 + np.sqrt(3) * d / self.rho)
+            * np.exp(-np.sqrt(3) * d / self.rho)
+        )
+
+    def _gradient(self, lags: np.ndarray):
+        raise NotImplementedError()
+
+
+class Matern52Model(CovarianceModel):
+    """
+    Implements the Matern Covariance kernel with slope parameter 5/2.
+
+    Attributes
+    ----------
+    rho: float
+        length scale parameter of the kernel
+
+    sigma: float
+        amplitude parameter of the kernel
+    """
+
+    rho = ModelParameter(default=1.0, bounds=(0, np.infty))
+    sigma = ModelParameter(default=1.0, bounds=(0, np.infty))
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def _compute(self, lags: np.ndarray):
+        d = np.sqrt(np.sum(lags**2, 0))
+        temp = np.sqrt(5) * d / self.rho
+        return self.sigma**2 * (1 + temp + temp**2 / 3) * np.exp(-temp)
+
+    def _gradient(self, lags: np.ndarray):
+        raise NotImplementedError()
+
+
+class RationalQuadraticModel(CovarianceModel):
+    """
+    Implements the Rational Quadratic Covariance Kernel.
+
+    Attributes
+    ----------
+    rho: float
+        length scale parameter of the kernel
+
+    alpha: float
+        alpha parameter of the kernel
+
+    sigma: float
+        amplitude parameter of the kernel
+    """
+
+    rho = ModelParameter(default=1.0, bounds=(0.0, np.inf))
+    alpha = ModelParameter(default=1.0, bounds=(0, np.inf))
+    sigma = ModelParameter(default=1.0, bounds=(0, np.inf))
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def _compute(self, lags: np.array):
+        d2 = np.sum(lags**2, 0) / (2 * self.rho**2)
+        return self.sigma**2 * np.power(1 + d2 / self.alpha, -self.alpha)
+
+    def _gradient(self, lags: np.ndarray):
+        raise NotImplementedError()
+
+
 class NuggetModel(CompoundModel):
     """
[…]
