+++ docs/models.md
@@ -6,7 +6,13 @@ to inherit from CovarianceModel or CompoundCovarianceModel.
 In both cases, you need to define the _compute method, which will expect
 a lag array with shape (ndim, n1, ..., nk, 1) where ndim is the number
 of spatio-temporal dimensions, n1, ..., nk, can have any shape.
-The trailing 1 is not needed in calls to the model as it is automatically added.
+The trailing 1 is not needed in calls to the model as it is automatically added by
+the inherited __call__ method.
+More specifically, the trailing dimension is used to adress the case where
+the user sets each parameter to be array-valued rather than scalar. This allows
+parallel evaluation of the likelihood for several model parameters. This feature
+can be ignored by new users of the package.
+
 Within the _compute method of a CompoundModel, when evaluating children models at lags,
 you should not use their call method but rather their _compute method (see e.g. SumModel).
 
@@ -20,5 +26,13 @@ you should not use their call method but rather their _compute method (see e.g. 
         - _gradient
 
 ::: debiased_spatial_whittle.models.SquaredExponentialModel
+
+::: debiased_spatial_whittle.models.Mater32Model
+
+::: debiased_spatial_whittle.models.Matern52Model
+
+::: debiased_spatial_whittle.models.RationalQuadraticModel
+
+::: debiased_spatial_whittle.models.AnisotropicModel
 
 ::: debiased_spatial_whittle.models.BivariateUniformCorrelation
