### 2025-03-20 78c4fc1
update docs with new example (anisotropic)
