### 2025-03-20 0ab760f
minor fixes for docs examples
