+++ docs/anisotropic.py
@@ -19,14 +19,16 @@ from debiased_spatial_whittle.likelihood import DebiasedWhittle, Estimator
 
 import matplotlib.pyplot as plt
 
-# ## Set up grid and model
+# ##Set up grid and model
+
 grid = RectangularGrid((256, 256))
 model = SquaredExponentialModel(rho=12.0)
 model = AnisotropicModel(model, eta=1.7, phi=np.pi / 4)
 model = NuggetModel(model, nugget=1e-2)
 model
 
-# ## Sample from the model
+# ##Sample from the model
+
 sampler = SamplerOnRectangularGrid(model, grid)
 data = sampler()
 
@@ -34,7 +36,8 @@ plt.figure()
 plt.pcolor(data, cmap="RdBu")
 plt.show()
 
-# ## Set up estimation
+# ##Set up estimation
+
 periodogram = Periodogram()
 expected_periodogram = ExpectedPeriodogram(grid, periodogram)
 
@@ -47,13 +50,15 @@ model_est_ = NuggetModel(model_est, nugget=1e-3)
 
 debiased_whittle = DebiasedWhittle(periodogram, expected_periodogram)
 
-# ## Carry out least square fit
+# ##Carry out least square fit
+
 least_square = LeastSquareEstimator(periodogram, expected_periodogram)
 least_square(data, model_est_)
 print(model_est_.free_parameter_values_to_array_deep())
 model_est_
 
-# ## Carry out Debiased Whittle fit
+# ##Carry out Debiased Whittle fit
+
 estimator = Estimator(debiased_whittle)
 estimator(model_est_, data)
 print(model_est_.free_parameter_values_to_array_deep())
