+++ docs/multivariate_two_grids.py
@@ -22,7 +22,7 @@ from debiased_spatial_whittle.simulation import SamplerBUCOnRectangularGrid
 
 # ##Grid specification
 
-g = RectangularGrid((128, 128), nvars=2)
+g = RectangularGrid((256, 256), nvars=2)
 g.mask = np.random.rand(*g.mask.shape) > 0.2
 x, y = g.grid_points
 g.mask[..., 0] = np.mod(x, 40) <= 20
