### 2025-03-22 94fffc3
simplified sum model and nugget model, add docs
