+++ src/debiased_spatial_whittle/models.py
@@ -298,6 +298,9 @@ class CovarianceModel(ModelInterface):
             out = np.squeeze(out, -1)
         return out
 
+    def __add__(self, other):
+        return SumModel(self, other)
+
 
 class CompoundModel(ModelInterface):
     def __init__(self, children, *args, **kwargs):
@@ -360,30 +363,59 @@ class CompoundModel(ModelInterface):
             out = np.squeeze(out, -1)
         return out
 
+    def __add__(self, other):
+        return SumModel(self, other)
+
 
 class SumModel(CompoundModel):
-    """Class that allows to define a new model as the sum of several models."""
+    """
+    Implements a covariance model defined as the sum of two several covariance models.
 
-    sigma = ModelParameter(default=1.0, bounds=(0, numpy.inf))
+    Examples
+    --------
+    >>> model_1 = SquaredExponentialModel(rho=32)
+    >>> model_2 = ExponentialModel(rho=5)
+    >>> model = model_1 + model_2
+    """
 
-    def __init__(self, children, *args, **kwargs):
-        super().__init__(children, *args, **kwargs)
+    def __new__(cls, *args, **kwargs):
+        children = []
+        for child in args:
+            if isinstance(child, SumModel):
+                children.append(child.children)
+            else:
+                children.append(child)
+        return super(SumModel, cls).__new__(cls)
+
+    def __init__(self, *args, **kwargs):
+        children = args
+        super().__init__(children, **kwargs)
 
     def _compute(self, lags: np.ndarray):
         values = (child._compute(lags) for child in self.children)
         out = sum(values)
-        return out / self._norm_constant() * self.sigma**2
-
-    def _norm_constant(self):
-        try:
-            sigmas = np.stack([child.sigma for child in self.children])
-        except TypeError:
-            sigmas = np.array([child.sigma for child in self.children])
-        out = np.sum(sigmas**2, axis=0)
         return out
 
 
 class ExponentialModel(CovarianceModel):
+    """
+    Implements the Exponential covariance model.
+
+    Attributes
+    ----------
+    rho: float
+        length scale parameter
+
+    sigma: float
+        amplitude parameter
+
+    Examples
+    --------
+    >>> model = ExponentialModel(rho=5, sigma=1.41)
+    >>> model(np.array([[0., 1.], [0., 0.]]))
+    array([1.9881    , 1.62771861])
+    """
+
     rho = ModelParameter(default=1.0, bounds=(0, numpy.inf), doc="Range parameter")
     sigma = ModelParameter(
         default=1.0, bounds=(0, numpy.inf), doc="Amplitude parameter"
@@ -411,6 +443,12 @@ class SquaredExponentialModel(CovarianceModel):
 
     sigma: float
         amplitude parameter
+
+    Examples
+    --------
+    >>> model = SquaredExponentialModel(rho=5, sigma=1.41)
+    >>> model(np.array([[0., 1.], [0., 0.]]))
+    array([1.9881    , 1.94873298])
     """
 
     rho = ModelParameter(default=1.0, bounds=(0, np.inf), doc="Range parameter")
@@ -453,6 +491,10 @@ class Matern32Model(CovarianceModel):
 
     sigma: float
         amplitude parameter of the kernel
+
+    Examples
+    --------
+    >>> model = Matern32Model(rho=5, sigma=1)
     """
 
     rho = ModelParameter(default=1.0, bounds=(0, np.inf))
@@ -484,6 +526,11 @@ class Matern52Model(CovarianceModel):
 
     sigma: float
         amplitude parameter of the kernel
+
+    Examples
+    --------
+    >>> model = Matern52Model(rho=10)
[…]
