### 2025-03-22 55c0f68
update residuals used in least squares
