### 2025-03-22 5ba27b0
New covariance models and least square estimator (#11)
* first version of least squares

* fix nugget model, fix least square for cupy

* add anisotropic model

* add matern32, matern52 and rational quadratic covariance kernels

* update model docs

* update docs with new example (anisotropic)

* update from actions/cache@v2 to actions/cache@v3 required

* fix typo

* minor fixes in docs

* minor fixes for docs examples

* add class MultivariateSamplerOnRectangularGrid that implements circulant embedding in multivaritate case

* simplified sum model and nugget model, add docs

* minor add to d
[…]
