### 2025-04-03 b0e9f15
allow some parameters to be optimized on log scale
