### 2025-04-03 1a9a617
set some parameters for log scale optimization by default
