### 2025-04-04 344187f
Log scale optim (#12)
* allow some parameters to be optimized on log scale

* set some parameters for log scale optimization by default
