### 2025-04-24 #7902 @arthurBarthe
Here is a list (order is irrelevant):

1. Marine Guyot  [MarineChap](https://github.com/MarineChap)
2. Robbi Bishop-Taylor  [robbibt](https://github.com/robbibt)
3. Natasha Yamane [nmy2103](https://github.com/nmy2103)
4. Anne Fouilloux [annefou](https://github.com/annefou)
5. Wei Ji Leong [weiji14](https://github.com/weiji14)
