### 2025-05-23 6fe02af
first version time series with long-memory
