### 2025-05-31 #8323 @editorialbot
[REVIEW]: DSWL: a Python implementation of the Debiased Spatial Whittle Likelihood
**Submitting author:** @ArthurBarthe (<a href="http://orcid.org/0000-0003-1571-4228">Arthur P. Guillaumin</a>)
**Repository:** https://github.com/arthurBarthe/debiased-spatial-whittle
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
