### 2025-06-16 arthurBarthe/debiased-spatial-whittle#13 @MarineChap
Import error from debiased_spatial_whittle.bayes
I tried to run example python file from the doc folder and the interpolation_example has an import error with a module bayes not found.
I can see the folder is in the .gitignore. What is the expected way to run interpolation in your package ?

https://github.com/openjournals/joss-reviews/issues/8323

Thank you.
