### 2025-06-16 arthurBarthe/debiased-spatial-whittle#14 @MarineChap
import error in circle_non_parametric.py and francebernouilli.py
Dear,

`ImportError: cannot import name 'AliasedSpectralModel' from 'debiased_spatial_whittle.models' `
Does this model have been replaced by something else ?

Also, there is probably an attribute name to change in francebernoulli.py

`AttributeError: 'SquaredExponentialModel' object has no attribute 'get_free_parameters_deep'. Did you mean: 'n_free_parameters_deep'?`

https://github.com/openjournals/joss-reviews/issues/8323

Have a good evening
