### 2025-06-16 arthurBarthe/debiased-spatial-whittle#15 @MarineChap
Pytest suite broken
Dear,
Few tests seem to be broken.  See pytest report below.
Most errors are due to a shape issue in debiased_spatial_whittle/models.py:769
Also, why do you test the version of the package in test_debiasedwhittle.py ?

https://github.com/openjournals/joss-reviews/issues/8323
Have a good evening

```
================================================================== test session starts ===================================================================
rootdir: /home/marine/Documents/debiased-spatial-whittle
configfile: pyproject.toml
collected 47 items

test_confidence.py .                                                                                                                               [  2%]
test_debiasedwhittle.py F                                                                                                                          [  4%]
test_estimation.py ...                                                                                                                             [ 10%]
test_expected_periodogram_multivariate.py F                                                                                                        [ 12%]
test_expected_periodogram.py ............F.....                                                                                                    [ 51%]
test_grids.py .                                                                                                                                    [ 53%]
test_hypothesis_tests.py F                                                                                                                         [ 55%]
test_likelihood.py ...F.F...F                                                                                                                      [ 76%]
test_models.py ....F..                                                                                                                             [ 91%]
test_simulation.py ....                                                                                                                            [100%]

======================================================================== FAILURES ========================================================================
______________________________________________________________________ test_version ______________________________________________________________________

    def test_version():
>       assert __version__ == "0.2"
E       AssertionError: assert '1.1.1' == '0.2'
E
E         - 0.2
E         + 1.1.1

test_debiasedwhittle.py:5: AssertionError
__________________________________________________ TestEpFullGrid.test_compare_to_average_periodiogram ___________________________________________________

self = <tests.test_expected_periodogram_multivariate.TestEpFullGrid object at 0x7f2442bb9660>

    def test_compare_to_average_periodiogram(self):
        sampler = SamplerBUCOnRectangularGrid(self.bvm, self.grid)
        p_computer = Periodogram()
        periodograms = []
        for i in range(10000):
            sample_i = sampler()
            p_value = p_computer([sample_i[..., 0], sample_i[..., 1]])
            periodograms.append(p_value)
        avg_per = np.mean(periodograms, axis=0)
        ep_computer = ExpectedPeriodogram(self.grid, p_computer)
        ep = ep_computer(self.bvm)
        # for this model and grid (full grid), the expected periodogram is real-valued
>       assert_allclose(np.real(avg_per), np.real(ep), rtol=0.2)

test_expected_periodogram_multivariate.py:30:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

args = (<function assert_allclose.<locals>.compare at 0x7f240d16ac20>, array([[[[ 91.57532559,  68.21956291],
         [ 68.2... [ 47.30557645, 394.21313711]]],

        [[[ 78.47011758,  58.85258818],
          [ 58.85258818, 490.43823484]]]]]))
kwds = {'equal_nan': True, 'err_msg': '', 'header': 'Not equal to tolerance rtol=0.2, atol=0', 'verbose': True}

    @wraps(func)
    def inner(*args, **kwds):
        with self._recreate_cm():
>           return func(*args, **kwds)
E           AssertionError:
E           Not equal to tolerance rtol=0.2, atol=0
E
E           (shapes (64, 64, 2, 2), (64, 64, 1, 2, 2) mismatch)
E            x: array([[[[ 91.575326,  68.219563],
E                    [ 68.219563, 575.711181]],
E           ...
E            y: array([[[[[ 90.804232,  68.103174],
E                     [ 68.103174, 567.526451]]],
E           ...

../../../micromamba/envs/debiased/lib/python3.10/contextlib.py:79: AssertionError
______________________________________________________ test_gradient_expected_periodogram_bivariate ______________________________________________________

    def test_gradient_expected_periodogram_bivariate():
        g = RectangularGrid((32, 32), nvars=2)
        p = PeriodogramMulti()
        ep_op = ExpectedPeriodogram(g, p)
        model = ExponentialModel(rho=3, sigma=1)
        bvm = BivariateUniformCorrelation(model)
        bvm.r = 0.1
        bvm.f = 1.2
        params_for_grad = [bvm.param.r, bvm.param.f]
>       ep_grad = ep_op.gradient(bvm, params_for_grad)

test_expected_periodogram.py:279:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
../src/debiased_spatial_whittle/periodogram.py:509: in gradient
    d_acv = model.gradient(lags, params)
../src/debiased_spatial_whittle/models.py:217: in gradient
    grad = self._gradient(lags)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = BivariateUniformCorrelation(f=1.2, free_only=True, name='BivariateUniformCorrelation00025', r=0.1)
x = array([[[-31., -31., -31., ..., -31., -31., -31.],
        [-30., -30., -30., ..., -30., -30., -30.],
        [-29., -....,  30.,  31.],
        [-31., -30., -29., ...,  29.,  30.,  31.],
        [-31., -30., -29., ...,  29.,  30.,  31.]]])

    def _gradient(self, x: np.ndarray):
        """

        Parameters
        ----------
        x
            shape (ndim, m1, ..., mk)

        Returns
        -------
        gradient
            shape (m1, ..., mk, 2, 2, p + 2)
            where p is the number of parameters of the base model.
        """
        acv11 = self.base_model(x)
        gradient_base_model = self.base_model._gradient(x)
        # gradient 11
        temp = np.stack((np.zeros_like(acv11), np.zeros_like(acv11)), axis=-1)
>       gradient_11 = np.concatenate((temp, gradient_base_model), axis=-1)
E       ValueError: all the input arrays must have same number of dimensions, but the array at index 0 has 3 dimension(s) and the array at index 1 has 0 dimension(s)

../src/debiased_spatial_whittle/models.py:769: ValueError
________________________________________________________________ test_fixed_parameters_ht ________________________________________________________________

    def test_fixed_parameters_ht():
        grid = RectangularGrid((64, 64), nvars=2)
        model = ExponentialModel()
        model.rho = 1
        model.sigma = 1
        model.nugget = 0.01
        bvm = BivariateUniformCorrelation(model)
        bvm.r_0 = 0.0
        bvm.f_0 = 1.0
        sampler = SamplerBUCOnRectangularGrid(bvm, grid)
        z = sampler()

        model.rho = None
        model.sigma = None
        bvm.r_0 = None
        bvm.f_0 = None

        periodogram = PeriodogramMulti()
        expected_periodogram = ExpectedPeriodogram(grid, periodogram)
        dbw = MultivariateDebiasedWhittle(periodogram, expected_periodogram)
        hypothesis_test = FixedParametersHT(bvm, dict(r_0=0.0), dbw)
>       hypothesis_test(z)

test_hypothesis_tests.py:36:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <debiased_spatial_whittle.hypothesis_tests.FixedParametersHT object at 0x7f240d19a2f0>
z = array([[[ 0.6841293 ,  0.52158469],
        [ 1.2037241 ,  0.39270845],
        [ 1.35359922,  0.43394988],
        .....     ...,
        [-1.82433731,  0.66178399],
        [-0.02497567,  0.21588331],
        [-0.30955543,  1.71677494]]])
level = 0.05

    def __call__(self, z: np.array, level=0.05):
>       params = self.full_model.free_params
E       AttributeError: 'BivariateUniformCorrelation' object has no attribute 'free_params'. Did you mean: 'free_parameters'?

../src/debiased_spatial_whittle/hypothesis_tests.py:63: AttributeError
________________________________________________________________ test_whittle_grad_multi _________________________________________________________________

    def test_whittle_grad_multi():
        """
        Tests the implementation of the gradient of the whittle likelihood in the multivariate case
        """
        g = RectangularGrid((32, 32), nvars=2)
        p = PeriodogramMulti()
        ep_op = ExpectedPeriodogram(g, p)
        model = SquaredExponentialModel(rho=3, sigma=1)
        bvm = BivariateUniformCorrelation(model)
        bvm.r = 0.3
        bvm.f = 1.5
        sampler = SamplerBUCOnRectangularGrid(bvm, g)
        z = sampler()
        dbw = MultivariateDebiasedWhittle(p, ep_op)
        epsilon = 1e-8
        params_for_grad = [bvm.param.r, bvm.param.f]
>       lkh, grad = dbw(z, bvm, params_for_grad)

test_likelihood.py:120:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
../src/debiased_spatial_whittle/likelihood.py:203: in __call__
    d_ep = self.expected_periodogram.gradient(model, params_for_gradient)
../src/debiased_spatial_whittle/periodogram.py:509: in gradient
    d_acv = model.gradient(lags, params)
../src/debiased_spatial_whittle/models.py:217: in gradient
    grad = self._gradient(lags)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = BivariateUniformCorrelation(f=1.5, free_only=True, name='BivariateUniformCorrelation00038', r=0.3)
x = array([[[-31., -31., -31., ..., -31., -31., -31.],
        [-30., -30., -30., ..., -30., -30., -30.],
        [-29., -....,  30.,  31.],
        [-31., -30., -29., ...,  29.,  30.,  31.],
        [-31., -30., -29., ...,  29.,  30.,  31.]]])

    def _gradient(self, x: np.ndarray):
        """

        Parameters
        ----------
        x
            shape (ndim, m1, ..., mk)

        Returns
        -------
        gradient
            shape (m1, ..., mk, 2, 2, p + 2)
            where p is the number of parameters of the base model.
        """
        acv11 = self.base_model(x)
        gradient_base_model = self.base_model._gradient(x)
        # gradient 11
        temp = np.stack((np.zeros_like(acv11), np.zeros_like(acv11)), axis=-1)
>       gradient_11 = np.concatenate((temp, gradient_base_model), axis=-1)
E       ValueError: all the input arrays must have same number of dimensions, but the array at index 0 has 3 dimension(s) and the array at index 1 has 0 dimension(s)

../src/debiased_spatial_whittle/models.py:769: ValueError
________________________________________________________________ test_fisher_multivariate ________________________________________________________________

    def test_fisher_multivariate():
        """
        Runs the fisher method in the multivariate case. Checks that the diagonal of the result is positive.
        """
        g = RectangularGrid((32, 32), nvars=2)
        p = PeriodogramMulti()
        ep_op = ExpectedPeriodogram(g, p)
        model = SquaredExponentialModel()
        model.rho = 3
        model.sigma = 1
        model.nugget = 0.2
        bvm = BivariateUniformCorrelation(model)
        bvm.r = 0.3
        bvm.f = 1.5
        sampler = SamplerBUCOnRectangularGrid(bvm, g)
        dbw = MultivariateDebiasedWhittle(p, ep_op)
>       h = dbw.fisher(bvm, [bvm.param.r, bvm.param.f])

test_likelihood.py:172:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
../src/debiased_spatial_whittle/likelihood.py:220: in fisher
    d_ep = self.expected_periodogram.gradient(model, params_for_gradient)
../src/debiased_spatial_whittle/periodogram.py:509: in gradient
    d_acv = model.gradient(lags, params)
../src/debiased_spatial_whittle/models.py:217: in gradient
    grad = self._gradient(lags)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = BivariateUniformCorrelation(f=1.5, free_only=True, name='BivariateUniformCorrelation00041', r=0.3)
x = array([[[-31., -31., -31., ..., -31., -31., -31.],
        [-30., -30., -30., ..., -30., -30., -30.],
        [-29., -....,  30.,  31.],
        [-31., -30., -29., ...,  29.,  30.,  31.],
        [-31., -30., -29., ...,  29.,  30.,  31.]]])

    def _gradient(self, x: np.ndarray):
        """

        Parameters
        ----------
        x
            shape (ndim, m1, ..., mk)

        Returns
        -------
        gradient
            shape (m1, ..., mk, 2, 2, p + 2)
            where p is the number of parameters of the base model.
        """
        acv11 = self.base_model(x)
        gradient_base_model = self.base_model._gradient(x)
        # gradient 11
        temp = np.stack((np.zeros_like(acv11), np.zeros_like(acv11)), axis=-1)
>       gradient_11 = np.concatenate((temp, gradient_base_model), axis=-1)
E       ValueError: all the input arrays must have same number of dimensions, but the array at index 0 has 3 dimension(s) and the array at index 1 has 0 dimension(s)

../src/debiased_spatial_whittle/models.py:769: ValueError
____________________________________________________________ test_jmatrix_sample_multivariate ____________________________________________________________

    def test_jmatrix_sample_multivariate():
        g = RectangularGrid((32, 32), nvars=2)
        p = PeriodogramMulti()
        ep_op = ExpectedPeriodogram(g, p)
        model = SquaredExponentialModel()
        model.rho = 3
        model.sigma = 1
        model.nugget = 0.2
        bvm = BivariateUniformCorrelation(model)
        bvm.r = 0.3
        bvm.f = 1.5
        dbw = MultivariateDebiasedWhittle(p, ep_op)
>       jmat = dbw.jmatrix_sample(bvm, [bvm.param.r, bvm.param.f])

test_likelihood.py:247:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
../src/debiased_spatial_whittle/likelihood.py:269: in jmatrix_sample
    _, grad = self(z, model, params_for_gradient)
../src/debiased_spatial_whittle/likelihood.py:203: in __call__
    d_ep = self.expected_periodogram.gradient(model, params_for_gradient)
../src/debiased_spatial_whittle/periodogram.py:509: in gradient
    d_acv = model.gradient(lags, params)
../src/debiased_spatial_whittle/models.py:217: in gradient
    grad = self._gradient(lags)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = BivariateUniformCorrelation(f=1.5, free_only=True, name='BivariateUniformCorrelation00046', r=0.3)
x = array([[[-31., -31., -31., ..., -31., -31., -31.],
        [-30., -30., -30., ..., -30., -30., -30.],
        [-29., -....,  30.,  31.],
        [-31., -30., -29., ...,  29.,  30.,  31.],
        [-31., -30., -29., ...,  29.,  30.,  31.]]])

    def _gradient(self, x: np.ndarray):
        """

        Parameters
        ----------
        x
            shape (ndim, m1, ..., mk)

        Returns
        -------
        gradient
            shape (m1, ..., mk, 2, 2, p + 2)
            where p is the number of parameters of the base model.
        """
        acv11 = self.base_model(x)
        gradient_base_model = self.base_model._gradient(x)
        # gradient 11
        temp = np.stack((np.zeros_like(acv11), np.zeros_like(acv11)), axis=-1)
>       gradient_11 = np.concatenate((temp, gradient_base_model), axis=-1)
E       ValueError: all the input arrays must have same number of dimensions, but the array at index 0 has 3 dimension(s) and the array at index 1 has 0 dimension(s)

../src/debiased_spatial_whittle/models.py:769: ValueError
________________________________________________________________ test_gradient_bivariate _________________________________________________________________

    def test_gradient_bivariate():
        """
        This test checks that the gradient has the right shape in the case of a bivariate model.
        Returns
        -------

        """
        g = RectangularGrid((32, 32), nvars=2)
        model = SquaredExponentialModel()
        model.rho = 3
        model.sigma = 2
        model.nugget = 0.2
        bvm = BivariateUniformCorrelation(model)
        bvm.r = 0.2
        bvm.f = 1.5
        lags = g.lags_unique
>       gradient = bvm.gradient(lags, [bvm.param.r, bvm.param.f])

test_models.py:90:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
../src/debiased_spatial_whittle/models.py:217: in gradient
    grad = self._gradient(lags)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = BivariateUniformCorrelation(f=1.5, free_only=True, name='BivariateUniformCorrelation00052', r=0.2)
x = array([[[-31., -31., -31., ..., -31., -31., -31.],
        [-30., -30., -30., ..., -30., -30., -30.],
        [-29., -....,  30.,  31.],
        [-31., -30., -29., ...,  29.,  30.,  31.],
        [-31., -30., -29., ...,  29.,  30.,  31.]]])

    def _gradient(self, x: np.ndarray):
        """

        Parameters
        ----------
        x
            shape (ndim, m1, ..., mk)

        Returns
        -------
        gradient
            shape (m1, ..., mk, 2, 2, p + 2)
            where p is the number of parameters of the base model.
        """
        acv11 = self.base_model(x)
        gradient_base_model = self.base_model._gradient(x)
        # gradient 11
        temp = np.stack((np.zeros_like(acv11), np.zeros_like(acv11)), axis=-1)
>       gradient_11 = np.concatenate((temp, gradient_base_model), axis=-1)
E       ValueError: all the input arrays must have same number of dimensions, but the array at index 0 has 3 dimension(s) and the array at index 1 has 0 dimension(s)

../src/debiased_spatial_whittle/models.py:769: ValueError
==================================================================== warnings summary ====================================================================
tests/test_estimation.py::test_estimation_1d
  /home/marine/Documents/debiased-spatial-whittle/src/debiased_spatial_whittle/likelihood.py:379: RuntimeWarning: divide by zero encountered in log
    (np.log(expected_periodogram) + periodogram / expected_periodogram)

tests/test_estimation.py::test_estimation_1d
  /home/marine/Documents/debiased-spatial-whittle/src/debiased_spatial_whittle/likelihood.py:379: RuntimeWarning: divide by zero encountered in divide
    (np.log(expected_periodogram) + periodogram / expected_periodogram)

tests/test_estimation.py::test_estimation_1d
  /home/marine/Documents/debiased-spatial-whittle/src/debiased_spatial_whittle/likelihood.py:379: RuntimeWarning: invalid value encountered in add
    (np.log(expected_periodogram) + periodogram / expected_periodogram)

tests/test_estimation.py::test_estimation_1d
  /home/marine/Documents/debiased-spatial-whittle/src/debiased_spatial_whittle/models.py:425: RuntimeWarning: divide by zero encountered in divide
[…]
