### 2025-06-16 arthurBarthe/debiased-spatial-whittle#15 @MarineChap
Also, the doctests are all passing (except the one with bayes import missing already mentionned in #13 )
