### 2025-06-16 arthurBarthe/debiased-spatial-whittle#16 @MarineChap
Documentation
Readme and setup part of the readthedoc website :

- It would be helpful to update the README with installation instructions for GPU environments, using [gpu12] or [gpu11]. Also, currently, the documentation website still refers to the older gpu keyword.
- A short section explaining how to quickly run the test suite with pytest would be very appreciated, especially for contributors who want to validate their setup.
- Linking the ReadTheDocs documentation in the README or in the repository's "About" section could make it easier to discover.

Notebook example section :

- Would it be possible to convert the [3dexponential.py example](https://debiased-spatial-whittle.readthedocs.io/en/latest/3dexponential.py) into a Jupyter notebook? That would help make the example more interactive and accessible.
- The [Multivariate_simple example](https://debiased-spatial-whittle.readthedocs.io/en/latest/multivariate_simple.html) currently throws an error during inference at the end, might be worth looking into.
- A notebook showcasing the performance differences between the available backends would be really insightful. It could help users choose the most efficient setup for their use case.

I will stop my reviewing here for the moment. There is some great work here and it was a pleasure testing out your code.
Have a good evening,

https://github.com/openjournals/joss-reviews/issues/8323
