### 2025-06-18 0813461
add missing spectral models
