### 2025-06-18 05adc0d
Add GPU installation instructions to README.md
