+++ README.md
@@ -20,9 +20,11 @@ The SDW extends ideas from the Whittle likelihood and Debiased Whittle Likelihoo
 
 ## Installation instructions
 
+### CPU-only
+
 The package can be installed via one of the following methods.
 
-1. Via the use of Poetry ([https://python-poetry.org/](https://python-poetry.org/)), by running the following command:
+1. Via the use of [Poetry](https://python-poetry.org/), by running the following command:
 
    ```bash
    poetry add debiased-spatial-whittle
@@ -33,6 +35,30 @@ The package can be installed via one of the following methods.
     ```bash
     pip install debiased-spatial-whittle
     ```
+
+### GPU
+The Debiased Spatial Whittle likelihood relies on the Fast Fourier Transform (FFT) for computational efficiency.
+GPU implementations of the FFT provide additional computational efficiency (order x100) at almost no additional cost thanks to GPU implementations of the FFT algorithm.
+
+If you want to install with GPU dependencies (Cupy and Pytorch):
+
+1. You need an NVIDIA GPU
+2. You need to install the CUDA Toolkit. See for instance Cupy's [installation page](https://docs.cupy.dev/en/stable/install.html).
+3. You can install Cupy or pytorch yourself in your environment. Or you can specify an extra to poetry, e.g.
+
+   ```bash
+   poetry add debiased-spatial-whittle -E gpu12
+   ```
+   if you version of the CUDA toolkit is 12.* (use gpu11 if your version is 11.*)
+
+You can then switch to using e.g. Cupy instead of numpy as the backend via:
+
+   ```python
+    from debiased_spatial_whittle.backend import BackendManager
+    BackendManager.set_backend("cupy")
+   ```
+
+This should be run before any other import from the debiased_spatial_whittle package.
 
 ## Development
 
