### 2025-06-18 9f3021b
Add developement instructions
