### 2025-06-18 8bb3154
Fix notebook issues (3dexponential.py and multivariate_simple.py)
