+++ docs/multivariate_simple.py
@@ -77,5 +77,6 @@ plt.show()
 e = Estimator(db)
 m = ExponentialModel(rho=1, sigma=1)
 bvm = BivariateUniformCorrelation(m, r=0.0, f=1.0)
-e(bvm, data)
+bvm.set_param_bounds(dict(r=(-0.95, 0.95)))
+e(bvm, data, opt_callback=lambda *args, **kwargs: print(bvm.r))
 print(bvm.r)
