### 2025-06-18 c489269
Fixes remaining issue with multivariate_simple.py (sigma and f parameters going to zero during optimization)
