### 2025-06-30 arthurBarthe/debiased-spatial-whittle#20 @weiji14
Reproducing CPU vs GPU performance claims
Hi there, just starting the first round of my reviews at https://github.com/openjournals/joss-reviews/issues/8323#issuecomment-2975096058.

L53-56 of the [draft paper](https://github.com/openjournals/joss-papers/blob/a0abcd55fb6b78e78ca2f93a8e086f585619f297/joss.08323/10.21105.joss.08323.pdf) makes this claim:
> Finally, the user can switch between several backends, Numpy, Cupy and PyTorch. This allows to further benefit from computational gains via GPU implementations of the Fast Fourier Transform. This is shown in Figure 3 where we observed a ×100 speed-up with a GPU versus a CPU.

I ran some tests with my GPU (NVIDIA RTX A2000 8GB GPU) on the draft `cpu_vs_gpu.ipynb` notebook at https://github.com/arthurBarthe/debiased-spatial-whittle/pull/18/files#diff-caf378bc76ff4411d3d5619fdbd4a532c3fef3bba74a578a38987c6ffdd92d9f (which I realize still needs some work), but managed to make some tweaks to get this plot:

![Image](https://github.com/user-attachments/assets/c9cf2101-17b3-4be0-9bf8-8542af51c220)

There is still a speedup with the GPU (cupy) implementation over CPU (numpy), but less pronounced than the one in the draft preprint.

![](https://github.com/arthurBarthe/debiased-spatial-whittle/blob/v1.1.1/paper/times.jpeg?raw=true)

Could you report the model of the CPU and GPU you used to produce the results above? Also maybe report the versions of NumPy and CuPy that you are using. Would be good to at least have a frame of reference when running the benchmarks.
