### 2025-07-10 #8323 @gkthiruvathukal
@arthurBarthe Please respond by July 16, if you wish to proceed with your JOSS submission. Thanks!
