+++ docs/contribute.md
@@ -86,10 +86,14 @@ changes should not fall outside the main scope of the package.
 
 Before a PR can be accepted, the following criteria are required:
 1. All unit tests should pass
+
 2. New functionality should come with new unit tests
+
 3. Ruff checks and format should pass
+
 4. New functionality should come with docstrings and a corresponding section in the mkdocs
 documentation.
+
 5. If a PR addresses a specific issue, the latter should be referenced in the PR's description
 
 Depending on the type of contribution (patch, minor, major), the version will be
