### 2025-08-01 3c1a562
add details on mkdocs use for documentation building
