### 2025-08-01 96937d6
add interpolation and an example notebook
