### 2025-08-01 arthurBarthe/debiased-spatial-whittle#13 @arthurBarthe
Hi,

interpolation_example.py is from extension work I conducted with a collaborator. We decided to separate the code for both works for now, so I should have removed that file.

However, it makes sense to have an example where we use the fitted model for interpolation, so I have created an example in #18. You can check it in the new documentation, https://debiased-spatial-whittle--18.org.readthedocs.build/en/18/.
