### 2025-08-02 #8323 @gkthiruvathukal
In my own research group, @arthurBarthe, we _broadly_ define criteria for including co-authors and do not always require a code-related contribution.

Making good software is not just about writing code and making commits. Ideas/design and informing use-cases are also equally important.  In this case, it appears @thomas-goodwin has been a significant contributor--at least in the past--when it comes to these aspects.  As the cognizant editor, I believe he meets the criteria for inclusion, especially if you want to include him. 89 commits is a significant number, regardless of the code size.

Apart from you including him, Tom should indicate that he *wants* to be included.

Let me know if you have any questions.
