### 2025-08-04 #8323 @arthurBarthe
Thanks for the explanation @gkthiruvathukal , @thomas-goodwin has confirmed to me that he would be happy to be added as a co-author. I will update the paper accordingly.
