### 2025-08-04 1e4c933
Implement falling back to default backend when not available + list of available backends
