### 2025-08-04 7c08cb0
changes the axis where the gradient is stored for simplification
