+++ tests/test_expected_periodogram_multivariate.py
@@ -11,9 +11,7 @@ from debiased_spatial_whittle.multivariate_periodogram import Periodogram
 class TestEpFullGrid:
     grid = RectangularGrid((64, 64), nvars=2)
     base_model = SquaredExponentialModel(rho=4, sigma=1)
-    bvm = BivariateUniformCorrelation(base_model)
-    bvm.r = 0.3
-    bvm.f = 2.5
+    bvm = BivariateUniformCorrelation(base_model, r=0.3, f=2.5)
 
     def test_compare_to_average_periodiogram(self):
         sampler = SamplerBUCOnRectangularGrid(self.bvm, self.grid)
@@ -25,6 +23,6 @@ class TestEpFullGrid:
             periodograms.append(p_value)
         avg_per = np.mean(periodograms, axis=0)
         ep_computer = ExpectedPeriodogram(self.grid, p_computer)
-        ep = ep_computer(self.bvm)[:, :, 0, ...]
+        ep = ep_computer(self.bvm)
         # for this model and grid (full grid), the expected periodogram is real-valued
         assert_allclose(np.real(avg_per), np.real(ep), rtol=0.2)
