### 2025-08-05 #8323 @thomas-goodwin
Hi @gkthiruvathukal and @arthurBarthe,

Thank you very much for pointing that out, @gkthiruvathukal.

Happy to be an additional author, and thank you @arthurBarthe, for including me.

Thanks,
Tom
