### 2025-08-06 ad17f46
add more tools for gof and their documentation / examples
