+++ src/debiased_spatial_whittle/diagnostics.py
@@ -1,17 +1,115 @@
 from functools import cached_property
 import numpy as np
-from scipy.stats import chisquare
+from numpy.fft import fftshift
+from scipy.stats import chi2, chisquare, uniform, norm
 from matplotlib import pyplot as plt
+
 from debiased_spatial_whittle.models import CovarianceModel
 from debiased_spatial_whittle.grids import RectangularGrid
 from debiased_spatial_whittle.periodogram import Periodogram, ExpectedPeriodogram
 from debiased_spatial_whittle.likelihood import DebiasedWhittle, Estimator
 from debiased_spatial_whittle.simulation import SamplerOnRectangularGrid
+from debiased_spatial_whittle.utils import plot_fourier_values
+
+
+class GoodnessOfFitSimonsOlhede:
+    """
+    Class to carry out a goodness-of-fit analysis that relies on spectral residuals as defined in
+    Simons & Olhede 2013.
+    """
+
+    def __init__(
+        self, model: CovarianceModel, grid: RectangularGrid, sample: np.ndarray
+    ):
+        self.model = model
+        self.grid = grid
+        self.sample = sample
+        self.periodogram_computer = Periodogram()
+
+    def compute_residuals(self):
+        r"""
+        Compute spectral residuals.
+
+        Notes
+        -----
+        The residuals are defined according to,
+
+        $$
+            X(k) = \frac{I(\mathbf{k})}{\overline{I}(\mathbf{k}; \widehat{\boldsymbol{\theta}})},
+        $$
+
+        where $\widehat{\boldsymbol{\theta}}$ is the model's parameter estimates. Under the correct model,
+        these residuals are expected to (approximately) follow a chi-square distribution with two degrees of freedom
+        multiplied by a factor one half.
+
+        Returns
+        -------
+        residuals
+            Array of spectral residuals.
+        """
+        periodogram = self.periodogram_computer(self.sample)
+        ep = ExpectedPeriodogram(self.grid, self.periodogram_computer)(self.model)
+        residuals = periodogram / ep
+        return residuals
+
+    def plot(self):
+        """
+        Generates two plots regarding the distribution of the spectral-domain residuals. A Fourier-space
+        plot of the residual values, and a qq-plot against the approximate theoretical distribution.
+
+        Examples
+        --------
+        >>> from debiased_spatial_whittle.models import ExponentialModel, SquaredExponentialModel
+        >>> from debiased_spatial_whittle.grids import RectangularGrid
+        >>> from debiased_spatial_whittle.simulation import SamplerOnRectangularGrid
+        >>> model = ExponentialModel(rho=8)
+        >>> grid = RectangularGrid((128, 128))
+        >>> sample = SamplerOnRectangularGrid(model, grid)()
+        >>> gof = GoodnessOfFitSimonsOlhede(model, grid, sample)
+        >>> gof.plot()
+        >>> model = SquaredExponentialModel(rho=8)
+        >>> gof = GoodnessOfFitSimonsOlhede(model, grid, sample)
+        >>> gof.plot()
+        """
+        residuals = self.compute_residuals()
+        # spatial plot
+        fig = plt.figure()
+        ax = fig.add_subplot(1, 2, 1)
+        plot_fourier_values(self.grid, fftshift(residuals), ax=ax, vmin=0, vmax=6)
+        ax.set_title("Fourier residuals")
+        # qq plot
+        dist = chi2(df=2)
+        ps = np.linspace(0, 1, 500)
+        th_quantiles = dist.ppf(ps) / 2
+        emp_quantiles = np.quantile(residuals, ps)
+        ax = fig.add_subplot(1, 2, 2)
+        ax.plot(th_quantiles, emp_quantiles, "-*")
+        ax.plot(th_quantiles, th_quantiles)
+        ax.set_aspect("equal")
[…]
