+++ src/debiased_spatial_whittle/utils.py
@@ -38,12 +38,25 @@ def plot_fourier_values(
     -------
     axis
         pyplot axis
+
+    Examples
+    --------
+    >>> import numpy as np
+    >>> from numpy.fft import fftshift
+    >>> from debiased_spatial_whittle import *
+    >>> from debiased_spatial_whittle.models import Matern32Model
+    >>> grid = RectangularGrid((128, 128))
+    >>> model = Matern32Model(rho=12)
+    >>> ep = ExpectedPeriodogram(grid, Periodogram())(model)
+    >>> ax = plot_fourier_values(grid, 10 * np.log10(fftshift(ep)), plot_func="pcolor", ax=None)
+    >>> plt.show()
     """
     frequencies = grid.fourier_frequencies
     frequencies = fftshift(frequencies, (0, 1))
     if ax is None:
         fig = plt.figure()
         ax = fig.add_subplot()
+    kwargs.setdefault("cmap", "pink")
     if plot_func == "imshow":
         extent = (
             frequencies[0, 0, 0],
@@ -52,6 +65,11 @@ def plot_fourier_values(
             frequencies[0, -1, 1],
         )
         im = ax.imshow(values, extent=extent, *args, **kwargs)
+        plt.colorbar(im, ax=ax)
+    if plot_func == "contourf" or plot_func == "contour" or plot_func == "pcolor":
+        im = getattr(ax, plot_func)(
+            frequencies[..., 0], frequencies[..., 1], values, *args, **kwargs
+        )
         plt.colorbar(im, ax=ax)
     return ax
 
