+++ src/debiased_spatial_whittle/__init__.py
@@ -1,6 +1,5 @@
 __version__ = "1.0.1"
 
-from debiased_spatial_whittle.grids import RectangularGrid
-from debiased_spatial_whittle.periodogram import Periodogram, ExpectedPeriodogram
-from debiased_spatial_whittle.simulation import SamplerOnRectangularGrid
-from debiased_spatial_whittle.likelihood import DebiasedWhittle, Estimator
+import matplotlib.pyplot as plt
+
+plt.style.use("bmh")
