### 2025-08-07 #8323 @weiji14
Hi @gkthiruvathukal, I am currently away travelling, so not able to review this properly for a couple of weeks. That said, I do think this paper is on the right track, though I still see the issue of not having the GPU make/model listed (https://github.com/arthurBarthe/debiased-spatial-whittle/issues/20) still in an unresolved state. I will take a closer look when I return to my country after 18 August.
