### 2025-08-07 0be7a26
add unit test for shapes of arrays
