### 2025-08-07 766449a
update speedup claim in paper and cpu-gpu times figure
