### 2025-08-07 0c0652e
reset execute of notebooks to true
