### 2025-08-07 #8323 @MarineChap
I continued my review. The paper is making good progress. I made some comments on PRs and open a new issue.
I would think the next step is to finish the prs and merge them before continuing ?
The two biggest point for me were the GPU performance and the author contribution issue. I think the state of both issues will be resolved once the PRs are merged.
