### 2025-08-08 55018cf
update mkdocstrings-python options
