### 2025-08-08 0996e85
update installation instructions of docs
