+++ src/debiased_spatial_whittle/periodogram.py
@@ -1,16 +1,16 @@
 from .backend import BackendManager
 
-np = BackendManager.get_backend()
+xp = BackendManager.get_backend()
 
 from typing import Tuple
 
 from debiased_spatial_whittle.spatial_kernel import spatial_kernel
 from debiased_spatial_whittle.utils import prod_list
 
-fft = np.fft.fft
-fftn = np.fft.fftn
-ifftshift = np.fft.ifftshift
-ndarray = np.ndarray
+fft = xp.fft.fft
+fftn = xp.fft.fftn
+ifftshift = xp.fft.ifftshift
+ndarray = xp.ndarray
 
 
 def autocov(cov_func, shape):
@@ -18,7 +18,7 @@ def autocov(cov_func, shape):
     In d=1 the lags would be -(n-1)...n-1, but then a iffshit is applied so that the lags are
     0 ... n-1 -n+1 ... -1. This may look weird but it makes it easier to do the folding operation
     when computing the expecting periodogram"""
-    xs = np.meshgrid(*(np.arange(-n + 1, n) for n in shape), indexing="ij")
+    xs = xp.meshgrid(*(xp.arange(-n + 1, n) for n in shape), indexing="ij")
     if BackendManager.backend_name == "torch":
         # TODO this is a temporary solution, not ideal though
         xs = xs.to(device=BackendManager.device)
@@ -44,12 +44,12 @@ def compute_ep(
     # now we need to "fold"
     if fold:
         # TODO: make this autograd compatible for any d with any n's
-        result = np.zeros(shape, dtype=np.complex128)
+        result = xp.zeros(shape, dtype=xp.complex128)
 
         if n_dim == 1:
             for i in range(2):
                 res = cbar[i * shape[0] : (i + 1) * shape[0]]
-                result += np.pad(res, (i, 0), mode="constant")
+                result += xp.pad(res, (i, 0), mode="constant")
 
         elif n_dim == 2:
             for i in range(2):
@@ -58,7 +58,7 @@ def compute_ep(
                         i * shape[0] : (i + 1) * shape[0],
                         j * shape[1] : (j + 1) * shape[1],
                     ]
-                    result += np.pad(
+                    result += xp.pad(
                         res, ((i, 0), (j, 0)), mode="constant"
                     )  # autograd solution
 
@@ -71,17 +71,17 @@ def compute_ep(
                             j * shape[1] : (j + 1) * shape[1],
                             k * shape[2] : (k + 1) * shape[2],
                         ]
-                        result += np.pad(res, ((i, 0), (j, 0), (k, 0)), mode="constant")
+                        result += xp.pad(res, ((i, 0), (j, 0), (k, 0)), mode="constant")
 
     else:
         m, n = shape
-        result = np.zeros((2 * m, 2 * n))
+        result = xp.zeros((2 * m, 2 * n))
         result[:m, :n] = cbar[:m, :n]
         result[m + 1 :, :n] = cbar[m:, :n]
         result[m + 1 :, n + 1 :] = cbar[m:, n:]
         result[:m, n + 1 :] = cbar[:m, n:]
     # We take the real part of the fft only due to numerical precision, in theory this should be real
-    result = np.real(fftn(result))
+    result = xp.real(fftn(result))
     return result
 
 
@@ -101,8 +101,8 @@ def compute_ep_old(cov_func, grid, fold=True):
     cbar = cg * acv
     # now we need to "fold"
     if fold:
-        result = np.zeros_like(grid)
-        result = np.zeros_like(grid, dtype=np.complex128)
+        result = xp.zeros_like(grid)
+        result = xp.zeros_like(grid, dtype=xp.complex128)
         if n_dim == 1:
[…]

[…]
