### 2025-08-09 1aa171f
fix import scipy.special & replace np with xp
