### 2025-08-09 23358c6
fix SpectralMatern for torch backend
