### 2025-08-09 f8c6d1c
add pytest-coverage dependency
