### 2025-08-09 e9dee4e
add unit test for multivariate simulation
