+++ src/debiased_spatial_whittle/models.py
@@ -8,6 +8,7 @@ except:
     import numpy as np
 
 import numpy
+from numpy.linalg import inv
 import param
 from param import Parameterized
 from param.parameterized import _get_param_repr
@@ -793,6 +794,159 @@ class BivariateUniformCorrelation(CompoundModel):
         d_f[..., 0, 1] = acv_base_model * self.r
         d_f[..., 1, 0] = acv_base_model * self.r
         return dict(r=d_r, f=d_f)
+
+
+class DualRemoteSensing(CompoundModel):
+    """
+    Class for the covariance model corresponding to a univariate field measured via two channels.
+    """
+
+    sigma1 = ModelParameter(
+        default=0.1, bounds=(0.0, 1.0), doc="Noise amplitude of channel 1"
+    )
+    sigma2 = ModelParameter(
+        default=0.1, bounds=(0.0, 1.0), doc="Noise amplitude of channel 2"
+    )
+
+    def __init__(self, base_model: ModelInterface, *args, **kwargs):
+        super().__init__(
+            children=[
+                base_model,
+            ],
+            *args,
+            **kwargs,
+        )
+
+    @property
+    def base_model(self):
+        return self.children[0]
+
+    def compute11(self, lags: np.ndarray):
+        acv_latent = self.children[0]._compute(lags)
+        return acv_latent + self.sigma1 * np.all(lags == 0, 0)
+
+    def compute22(self, lags: np.ndarray):
+        acv_latent = self.children[0]._compute(lags)
+        return acv_latent + self.sigma2 * np.all(lags == 0, 0)
+
+    def _compute(self, lags: np.ndarray):
+        acv_latent = self.children[0]._compute(lags)
+        col1 = np.stack(
+            (acv_latent + self.sigma1 * np.all(lags == 0, 0), acv_latent), -1
+        )
+        col2 = np.stack(
+            (acv_latent, acv_latent + self.sigma2 * np.all(lags == 0, 0)), -1
+        )
+        return np.stack((col1, col2), -1)
+
+    def _gradient(self, lags: np.ndarray):
+        raise NotImplementedError()
+
+    def cov_mat_x1_x2(self, x1: np.ndarray, x2: np.ndarray = None) -> np.ndarray:
+        """
+        For multivariate data, if x1 has n1 locations and x2 has n2 locations, we return a matrix
+        with size (n1 * 2, n1 * 2)
+        Parameters
+        ----------
+        x1
+        x2
+
+        Returns
+        -------
+
+        """
+        raise NotImplementedError()
+        if x2 is None:
+            x2 = x1
+        x1 = np.expand_dims(x1, axis=1)
+        x2 = np.expand_dims(x2, axis=0)
+        lags = x1 - x2
+        lags = np.transpose(lags, (2, 0, 1))
+        acv = self(lags)
+        acv = np.transpose(acv, (3, 1, 4, 2))
+        return np.reshape(acv, (2 * x1.shape[0], 2 * x2.shape[0]))
+
+    def predict(
+        self,
+        x_obs: np.ndarray,
+        y_obs: np.ndarray,
+        x_pred: np.ndarray,
+        return_variance: bool = False,
+        full_model: ModelInterface = None,
+    ):
+        """
+        Compute conditional mean at a set of locations x_pred given values y_obs observed at x_obs.
+
+        Parameters
+        ----------
+        x_obs
+            shape (n_obs, d), array of locations where observations are made
+        y_obs
[…]
