### 2025-08-13 eec0ed8
initial version for jax backend (todo: random state)
