### 2025-08-28 #8323 @arthurBarthe
@gkthiruvathukal Sorry for the delay, and thank you for confirming. My co-authors have all agreed. I have updated the paper and added both Tom and Olivia as co-authors. I have created a [PR](https://github.com/arthurBarthe/debiased-spatial-whittle/pull/21) - although I just saw that I messed up the order of the affiliations, I will update that.
