### 2025-09-05 4e3bcf8
add dual remote sensing example to docs
