+++ docs/dual_remote_sensing.py
@@ -1,3 +1,11 @@
+# In this notebook we consider an application to dual remote sensing: a univariate field is remotely-sensed, with
+# each measurement having a white noise error of a certain level.
+# We therefore jointly estimate the noise level for each sensor in parallel with the parameters of the covariance model
+# of the sensed field.
+# We then proceed to "map" the field using both measurements.
+
+# ##Imports
+
 import matplotlib.pyplot as plt
 
 from debiased_spatial_whittle.models import SquaredExponentialModel, DualRemoteSensing
@@ -8,13 +16,16 @@ from debiased_spatial_whittle.multivariate_periodogram import Periodogram
 from debiased_spatial_whittle.periodogram import ExpectedPeriodogram
 from debiased_spatial_whittle.likelihood import MultivariateDebiasedWhittle, Estimator
 
+# ##Grid and model specification
+
 grid = RectangularGrid((64, 64), nvars=2)
 latent_model = SquaredExponentialModel(rho=5.0)
 model = DualRemoteSensing(latent_model, sigma1=0.1, sigma2=0.2)
 
+# ##Sample a realization
+
 sampler = MultivariateSamplerOnRectangularGrid(model, grid, p=2)
 s = sampler()
-print(s.shape)
 
 plt.figure()
 plt.subplot(1, 2, 1)
@@ -23,7 +34,8 @@ plt.subplot(1, 2, 2)
 plt.imshow(s[..., 1], cmap="seismic")
 plt.show()
 
-# inference
+# ##Parameter inference
+
 latent_model = SquaredExponentialModel(rho=1.0)
 model = DualRemoteSensing(latent_model, sigma1=0.01, sigma2=0.01)
 model.set_param_bounds(dict(sigma1=(0.001, 1.0), sigma2=(0.001, 1.0)))
@@ -33,7 +45,8 @@ dbw = MultivariateDebiasedWhittle(periodogram, ExpectedPeriodogram(grid, periodo
 estimator = Estimator(dbw)
 estimator(model, s, opt_callback=lambda *args, **kwargs: print(*args, **kwargs))
 
-# predict
+# ##Mapping
+
 import numpy as np
 
 xs = np.stack(np.mgrid[:64, :64], -1).reshape(-1, 2)
