### 2025-10-07 f069ff5
fix unit tests (#22)
* fix unit tests

* delete version test

* minor update

* changes the axis where the gradient is stored for simplification

* add unit test for shapes of arrays

* reset execute of notebooks to true

* minor fix

* add pytest-coverage dependency

* add unit test for multivariate simulation
