### 2025-10-07 2610f75
fix import when cupy not installed
