### 2025-10-07 98280ee
Fix documentation (#16)  (#18)
* Add GPU installation instructions to README.md

* Add developement instructions

* Add developement instructions

* Fix notebook issues (3dexponential.py and multivariate_simple.py)

* Fixes remaining issue with multivariate_simple.py (sigma and f parameters going to zero during optimization)

* First version of cpu vs gpu notebook

* add link to documentation website

* add CONTRIBUTING.md (symlink to docs/contributed.md)

* add details

* fix formatting issue

* add details on mkdocs use for documentation building

* minor fix

* add interpolation and an exam
[…]
