+++ src/debiased_spatial_whittle/new_models.py
@@ -1,410 +1,0 @@
-from abc import ABC, abstractmethod, abstractproperty
-import pickle
-from debiased_spatial_whittle.backend import BackendManager
-
-try:
-    np = BackendManager.get_backend()
-except:
-    import numpy as np
-import param
-from param import Parameterized
-from param.parameterized import _get_param_repr
-
-zeros = BackendManager.get_zeros()
-
-
-def _parameterized_repr_html(p, open):
-    """HTML representation for a Parameterized object"""
-    if isinstance(p, Parameterized):
-        cls = p.__class__
-        title = cls.name + "()"
-        value_field = "Value"
-    else:
-        cls = p
-        title = cls.name
-        value_field = "Default"
-
-    tooltip_css = """
-        .param-doc-tooltip{
-          position: relative;
-          cursor: help;
-        }
-        .param-doc-tooltip:hover:after{
-          content: attr(data-tooltip);
-          background-color: black;
-          color: #fff;
-          border-radius: 3px;
-          padding: 10px;
-          position: absolute;
-          z-index: 1;
-          top: -5px;
-          left: 100%;
-          margin-left: 10px;
-          min-width: 250px;
-        }
-        .param-doc-tooltip:hover:before {
-          content: "";
-          position: absolute;
-          top: 50%;
-          left: 100%;
-          margin-top: -5px;
-          border-width: 5px;
-          border-style: solid;
-          border-color: transparent black transparent transparent;
-        }
-        """
-    openstr = " open" if open else ""
-    param_values = p.param.values().items()
-    # contents = "".join(_get_param_repr(key, val, p.param[key])
-    #                   for key, val in param_values)
-    contents = ""
-    for key, val in param_values:
-        if key in ("name", "free_only"):
-            continue
-        if not p.param[key].readonly:
-            contents += _get_param_repr(key, val, p.param[key])
-        else:
-            contents += (
-                '<tr style="color:coral">'
-                + _get_param_repr(key, val, p.param[key])[4:-7]
-                + "<\tr>"
-            )
-    return (
-        f"<style>{tooltip_css}</style>\n"
-        f"<details {openstr}>\n"
-        ' <summary style="display:list-item; outline:none;">\n'
-        f"  <tt>{title}</tt>\n"
-        " </summary>\n"
-        ' <div style="padding-left:10px; padding-bottom:5px;">\n'
-        '  <table style="max-width:100%; border:1px solid #AAAAAA;">\n'
-        f'   <tr><th style="text-align:left;">Name</th><th style="text-align:left;">{value_field}</th><th style="text-align:left;">Type</th><th>Range</th></tr>\n'
-        f"{contents}\n"
-        "  </table>\n </div>\n</details>\n"
-    )
-
-
-class ModelParameter(param.Parameter):
-    __slots__ = [
-        "bounds",
-    ]
-
-    def __init__(self, *args, **kwargs):
-        self.bounds = kwargs.pop("bounds")
-        super().__init__(*args, allow_refs=True, per_instance=True, **kwargs)
-
-    @property
-    def free(self) -> bool:
-        """true is the model parameter is free, i.e. not readonly and not constant"""
-        return not (self.readonly or self.constant)
-
-
-class ModelInterface(param.Parameterized):
-    """
-    Class defining the general interface for covariance models
-
-    Attributes
-    ----------
-
-    """
-
-    free_only = param.Boolean(per_instance=True, default=True)
-
-    @abstractmethod
-    def __call__(self, lags: np.ndarray):
-        """
-        Evaluate the covariance model at the passed lags.
-
-        Parameters
-        ----------
-        lags
-            array of lags. Shape (ndim, n1, ..., nk) where ndim is the number of spatial dimensions.
-
-        Returns
-        -------
[…]
