### 2025-10-07 c302f55
Update paper with minor corrections + updated authorship (#21)
* minor corrections from sofia

* update author list
