### 2025-10-07 40a8508
add missing spectral models #14 (#17)
* add missing spectral models

* delete file circle_non_parametric.py

* fix import scipy.special & replace np with xp

* fix SpectralMatern for torch backend
