### 2025-10-07 43010f7
update reported specs to also include CPU
