### 2025-10-07 6dd0129
Minor fix - remove torch usage in docs
