### 2025-10-07 arthurBarthe/debiased-spatial-whittle#30 @weiji14
Complete API documentation
_Originally posted by @weiji14 in [#8323](https://github.com/openjournals/joss-reviews/issues/8323#issuecomment-2975096058)_

> - [ ] **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>> API documentation appears incomplete. E.g. at https://debiased-spatial-whittle.readthedocs.io/en/latest/grids.html, derived from https://github.com/arthurBarthe/debiased-spatial-whittle/blob/v1.1.1/docs/grids.md, only `debiased_spatial_whittle.grids.RectangularGrid` is documented, but there are many other grid classes at https://github.com/arthurBarthe/debiased-spatial-whittle/blob/v1.1.1/src/debiased_spatial_whittle/grids.py (e.g. [`CircleGrid`](https://github.com/arthurBarthe/debiased-spatial-whittle/blob/v1.1.1/src/debiased_spatial_whittle/grids.py#L50), [`ImgGrid`](https://github.com/arthurBarthe/debiased-spatial-whittle/blob/v1.1.1/src/debiased_spatial_whittle/grids.py#L84), etc). Recommend using something like sphinx autodoc's [automodule](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html#automatically-document-modules) directive to automatically document all classes/methods.

Not all classes are documented in the API reference page, e.g.:
- Grids (https://debiased-spatial-whittle.readthedocs.io/en/latest/grids.html) - missing `CircleGrid`, `BernoulliGrid`, `ImgGrid`, etc
- Models (https://github.com/arthurBarthe/debiased-spatial-whittle/blob/v1.1.1/docs/models.md) - missing `CompoundModel`
- Estimation (https://debiased-spatial-whittle.readthedocs.io/en/latest/periodogram.html) - missing `SeparableExpectedPeriodogram`

I'd recommend checking out resources such as https://www.pyopensci.org/python-package-guide/documentation/write-user-documentation/document-your-code-api-docstrings.html#document-the-code-in-your-package-s-api-using-docstrings and/or looking at how other packages do it (e.g. at https://www.fatiando.org/verde/latest/api/index.html). Let me know if you need any help on this, and would be happy to advise further!
