### 2025-10-08 21223e9
update speedup claim in paper and cpu-gpu times figure (#26)
* update speedup claim in paper and cpu-gpu times figure

* update reported specs to also include CPU
