### 2025-10-08 arthurBarthe/debiased-spatial-whittle#30 @arthurBarthe
Thanks, there are distinct reasons (not necessarily good ones!) for these classes not to be documented in the reference section of the documentation:
1. Grids: the mentioned classes  are only used in some example of the documentation. Maybe I should put them in a separate file to make it clearer that they are not core functionality?
2. Models: here Compound model should definitely be documented, I will make the update.
3. SeparableExpectedPeriodogram: this is currently not functional. I guess it's probably better to simply remove this class for now?

I use mkdocs to generate the documentation from the docstrings. But indeed, at the moment I specify manually which objects should appear in the API reference for each module in the documentation.

If that makes sense, I will clean up some of the files so that core / functional pieces are well separated from non-core / non-functional pieces. And if possible, I will change mkdocs settings to automatically document all classes.
