### 2025-10-09 fe6f2de
allow jax with nvidia gpu install
