### 2025-10-09 07b0a07
allow jax with nvidia gpu install
