+++ pyproject.toml
@@ -24,9 +24,11 @@ torch = {version="2.2.2", optional=true}
 jax = "^0.7.0"
 
 [tool.poetry.extras]
-gpu13 = ["cupy-cuda13x", "torch", "jax[cuda13]"]
 gpu12 = ["cupy-cuda12x", "torch", "jax[cuda12]"]
 gpu11 = ["cupy-cuda11x", "torch"]
+
+[tool.poetry.group.gpu13.dependencies]
+jax = { extras = ["cuda13"], version = "^0.4.20", platform = "linux" }
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.3.1"
