### 2025-10-28 7a539fb
JAX with cuda
