+++ docs/france_interpolation.py
@@ -9,11 +9,11 @@ from debiased_spatial_whittle.backend import BackendManager
 BackendManager.set_backend("numpy")
 import matplotlib.pyplot as plt
 import debiased_spatial_whittle.grids as grids
-from debiased_spatial_whittle.models import SquaredExponentialModel, NuggetModel
-from debiased_spatial_whittle.grids import RectangularGrid
-from debiased_spatial_whittle.simulation import SamplerOnRectangularGrid
-from debiased_spatial_whittle.periodogram import Periodogram, ExpectedPeriodogram
-from debiased_spatial_whittle.likelihood import Estimator, DebiasedWhittle
+from debiased_spatial_whittle.models.univariate import SquaredExponentialModel, NuggetModel
+from debiased_spatial_whittle.grids.base import RectangularGrid
+from debiased_spatial_whittle.sampling.simulation import SamplerOnRectangularGrid
+from debiased_spatial_whittle.inference.periodogram import Periodogram, ExpectedPeriodogram
+from debiased_spatial_whittle.inference.likelihood import Estimator, DebiasedWhittle
 
 
 # ##Model specification
