+++ docs/circle_gpu.py
@@ -11,11 +11,11 @@ np = BackendManager.get_backend()
 # ##Imports
 
 import matplotlib.pyplot as plt
-from debiased_spatial_whittle.models import SquaredExponentialModel
-from debiased_spatial_whittle.grids import RectangularGrid
-from debiased_spatial_whittle.simulation import SamplerOnRectangularGrid
-from debiased_spatial_whittle.periodogram import Periodogram, ExpectedPeriodogram
-from debiased_spatial_whittle.likelihood import Estimator, DebiasedWhittle
+from debiased_spatial_whittle.models.univariate import SquaredExponentialModel
+from debiased_spatial_whittle.grids.base import RectangularGrid
+from debiased_spatial_whittle.sampling.simulation import SamplerOnRectangularGrid
+from debiased_spatial_whittle.inference.periodogram import Periodogram, ExpectedPeriodogram
+from debiased_spatial_whittle.inference.likelihood import Estimator, DebiasedWhittle
 
 # ##Model Specification
 
