### 2025-10-30 e563918
fix execute to true in mkdocs.yaml
