+++ src/debiased_spatial_whittle/grids/old.py
@@ -15,7 +15,7 @@ ones = BackendManager.get_ones()
 fftfreq = np.fft.fftfreq
 from debiased_spatial_whittle.grids.spatial_kernel import spatial_kernel
 
-PATH_TO_FRANCE_IMG = str(Path(__file__).parents[2] / "france.jpg")
+PATH_TO_FRANCE_IMG = str(Path(__file__).parents[3] / "france.jpg")
 
 
 class Grid(ABC):
