+++ tests/test_estimation.py
@@ -3,7 +3,8 @@ from numpy.testing import assert_almost_equal
 from debiased_spatial_whittle.models.old import exp_cov
 from debiased_spatial_whittle.grids.base import RectangularGrid
 from debiased_spatial_whittle.inference.periodogram import Periodogram, ExpectedPeriodogram
-from debiased_spatial_whittle.inference.likelihood import DebiasedWhittle, Estimator, fit
+from debiased_spatial_whittle.inference.likelihood import DebiasedWhittle, Estimator
+from debiased_spatial_whittle.inference.old import fit
 from debiased_spatial_whittle.sampling.simulation import SamplerOnRectangularGrid
 from debiased_spatial_whittle.models.univariate import ExponentialModel
 
