+++ docs/sampling.md
@@ -1,0 +1,7 @@
+This package provides tools to efficiently sample from covariance models on
+grids using circulant embeddings and the Fast Fourier Transform.
+
+### ::: debiased_spatial_whittle.sampling.simulation
+
+### ::: debiased_spatial_whittle.sampling.samples
+
