+++ docs/grids.md
@@ -1,6 +1,3 @@
-## grids
-    options:
-      members: false
 This package provides tools to define sampling grids. Currently, the code
 is restricted to orthogonal grids. This is not a limitation inherent to the proposed
 method, and it would not be very hard to generalize to non-orthogonal grids.
