### 2025-11-14 c70c2f3
add missing docstrings
