### 2025-11-14 dbfd7b4
fix parameter type in docstrings
