### 2025-11-14 a34a958
Revamp overall structure + update API reference (#31)
* initial commit with new structure

* fix execute to true in mkdocs.yaml

* fix imports

* minor edits

* structure reference according to subpackages

* reset execute notebooks to true

* fix heading

* add missing docstrings

* fix parameter type in docstrings

---------

Co-authored-by: QMUL User Account <user@cln-msl-74vl3t3.qm.ds.qmul.ac.uk>
