### 2025-11-14 92252d4
Bump version from 1.1.1 to 2.1.1 - major
The overall structure of the package has been updated, by adding subpackages
