+++ src/debiased_spatial_whittle/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.1.1"
+__version__ = "2.1.1"
