### 2025-11-14 6516a6c
Update behaviour of BackendManager (#24)
* Implement falling back to default backend when not available + list of available backends

* fix import when cupy not installed

* minor fix

---------

Co-authored-by: QMUL User Account <user@cln-msl-74vl3t3.qm.ds.qmul.ac.uk>
