### 2025-11-14 #8323 @weiji14
Excellent, I see the new reference page is divided into [grids](https://debiased-spatial-whittle.readthedocs.io/en/latest/grids.html), [models](https://debiased-spatial-whittle.readthedocs.io/en/latest/models.html), [sampling](https://debiased-spatial-whittle.readthedocs.io/en/latest/sampling.html) and [inference](https://debiased-spatial-whittle.readthedocs.io/en/latest/estimation.html). This addresses the 'functionality' checkbox for me. (One optional tip I have is you can generate [different versions of your documentation](https://docs.readthedocs.com/platform/stable/versions.html) if you go to https://app.readthedocs.org/projects/debiased-spatial-whittle/builds/ and login to enable it).

I'm travelling for a conference soon, but once you complete the np -> xp rename, I can take a final look around end of November.
