+++ src/debiased_spatial_whittle/inference/multivariate_periodogram.py
@@ -1,8 +1,8 @@
 from debiased_spatial_whittle.backend import BackendManager
 
-np = BackendManager.get_backend()
+xp = BackendManager.get_backend()
 
-fftn = np.fft.fftn
+fftn = xp.fft.fftn
 
 from typing import List
 
@@ -15,9 +15,9 @@ class Periodogram:
     def __init__(self):
         # TODO allow for tapering in multivariate case
         self.fold = True
-        self.taper = lambda shape: np.ones(shape)
+        self.taper = lambda shape: xp.ones(shape)
 
-    def __call__(self, z: List[np.ndarray], return_fft: bool = False) -> np.ndarray:
+    def __call__(self, z: List[xp.ndarray], return_fft: bool = False) -> xp.ndarray:
         """
         Compute the multivariate periodogram. The data z is expected to be a list
         of p arrays with the same shape, where p is the number of variates.
@@ -36,20 +36,20 @@ class Periodogram:
             Shape (n1, n2, ..., nd, p, p) if the data is p-variate and over d spatial dimensions.
         """
         n_spatial_dims = z[0].ndim
-        z = np.stack(z, axis=-1)
+        z = xp.stack(z, axis=-1)
         # TODO generalize to 3d
         j_vec = (
-            1
-            / np.sqrt(np.array(z.shape[0] * z.shape[1]))
-            * fftn(z, None, list(range(n_spatial_dims)))
+                1
+                / xp.sqrt(xp.array(z.shape[0] * z.shape[1]))
+                * fftn(z, None, list(range(n_spatial_dims)))
         )
-        j_vec = np.expand_dims(j_vec, -1)
+        j_vec = xp.expand_dims(j_vec, -1)
         if return_fft:
             return j_vec
         # first dimensions are spatial dimensions
         if BackendManager.backend_name in ("numpy", "cupy"):
-            j_vec_transpose = np.conj(np.transpose(j_vec, (0, 1, -1, -2)))
+            j_vec_transpose = xp.conj(xp.transpose(j_vec, (0, 1, -1, -2)))
         elif BackendManager.backend_name == "torch":
-            j_vec_transpose = np.conj(np.transpose(j_vec, -1, -2))
-        p = np.matmul(j_vec, j_vec_transpose)
+            j_vec_transpose = xp.conj(xp.transpose(j_vec, -1, -2))
+        p = xp.matmul(j_vec, j_vec_transpose)
         return p
