+++ src/debiased_spatial_whittle/inference/confidence.py
@@ -1,6 +1,6 @@
 from debiased_spatial_whittle.backend import BackendManager
 
-np = BackendManager.get_backend()
+xp = BackendManager.get_backend()
 
 from progressbar import progressbar
 
@@ -21,46 +21,46 @@ class CovarianceFFT:
     def _get_indices_ep1(n, m):
         n1, n2 = n
         m1, m2 = m
-        x, y = np.meshgrid(np.arange(n1), np.arange(n2), indexing="ij")
-        a = np.logical_and(x < n1 - m1, y < n2 - m2)
-        b = np.logical_and(x >= -m1, y >= -m2)
-        return np.logical_and(a, b)
+        x, y = xp.meshgrid(xp.arange(n1), xp.arange(n2), indexing="ij")
+        a = xp.logical_and(x < n1 - m1, y < n2 - m2)
+        b = xp.logical_and(x >= -m1, y >= -m2)
+        return xp.logical_and(a, b)
 
     @staticmethod
     def _get_indices_ep2(n, m):
         n1, n2 = n
         m1, m2 = m
-        x, y = np.meshgrid(np.arange(n1), np.arange(n2), indexing="ij")
-        a = np.logical_and(x >= m1, y >= m2)
-        b = np.logical_and(x < n1 + m1, y < n2 + m2)
-        return np.logical_and(a, b)
+        x, y = xp.meshgrid(xp.arange(n1), xp.arange(n2), indexing="ij")
+        a = xp.logical_and(x >= m1, y >= m2)
+        b = xp.logical_and(x < n1 + m1, y < n2 + m2)
+        return xp.logical_and(a, b)
 
     @staticmethod
     def _get_indices_ep3(n, m):
         n1, n2 = n
         m1, m2 = m
-        x, y = np.meshgrid(np.arange(n1), np.arange(n2), indexing="ij")
-        a = np.logical_and(x <= m1, y <= m2)
-        b = np.logical_and(x >= m1 - n1 + 1, y >= m2 - n2 + 1)
-        return np.logical_and(a, b)
+        x, y = xp.meshgrid(xp.arange(n1), xp.arange(n2), indexing="ij")
+        a = xp.logical_and(x <= m1, y <= m2)
+        b = xp.logical_and(x >= m1 - n1 + 1, y >= m2 - n2 + 1)
+        return xp.logical_and(a, b)
 
     @staticmethod
     def _get_indices_ep4(n, m):
         n1, n2 = n
         m1, m2 = m
-        x, y = np.meshgrid(
-            np.arange(n1 - 1, -1, -1), np.arange(n2 - 1, -1, -1), indexing="ij"
+        x, y = xp.meshgrid(
+            xp.arange(n1 - 1, -1, -1), xp.arange(n2 - 1, -1, -1), indexing="ij"
         )
-        a = np.logical_and(x <= m1, y <= m2)
-        b = np.logical_and(x >= m1 - n1 + 1, y >= m2 - n2 + 1)
-        return np.logical_and(a, b)
+        a = xp.logical_and(x <= m1, y <= m2)
+        b = xp.logical_and(x >= m1 - n1 + 1, y >= m2 - n2 + 1)
+        return xp.logical_and(a, b)
 
     def exact_summation1(
         self,
         model: CovarianceModel,
         expected_periodogram: ExpectedPeriodogram,
-        f: np.ndarray = None,
-        f2: np.ndarray = None,
+        f: xp.ndarray = None,
+        f2: xp.ndarray = None,
         return_terms: bool = False,
         normalize: bool = False,
     ):
@@ -70,12 +70,12 @@ class CovarianceFFT:
         ep = expected_periodogram(model)
         s = []
         if f is None:
[…]

[…]
