### 2025-11-24 #8323 @arthurBarthe
Thanks for the feedback @weiji14 and for the remark regarding versioning of the documentation. I have added a version for the code before the restructuring (1.x) . I was also reading about having a developement branch, this could be useful in the future, I will consider it.

Additionally, I have now carried out the np -> xp rename and merged into the main branch.

@gkthiruvathukal I forgot to mention that due to the restructuring I carried out in  https://github.com/arthurBarthe/debiased-spatial-whittle/pull/31 I also did a major version bump (as most previous imports would break) to 2.x, whereas my initial submission to JOSS was for v1.0.1. Not sure if this has any impact at all, but I thought it might be better to notify you.
