### 2025-11-24 e8bd46e
List additional contributors in README
Added contributors to the README file.
