### 2025-11-27 arthurBarthe/debiased-spatial-whittle#34 @weiji14
Hardcoded path to france.jpg file
Hi there,

I'm reviewing https://github.com/openjournals/joss-reviews/issues/8323 once more with a fresh installation of [debiased-spatial-whittle v2.1.1](https://pypi.org/project/debiased-spatial-whittle/2.1.1), and running through the example at https://debiased-spatial-whittle.readthedocs.io/latest/france_interpolation.html#grid-specification, I'm getting the following error:

```python
In [1]: import numpy as np

In [2]: from debiased_spatial_whittle.backend import BackendManager

In [3]: BackendManager.set_backend("numpy")
   ...: import matplotlib.pyplot as plt
   ...: import debiased_spatial_whittle.grids as grids
   ...: from debiased_spatial_whittle.models.univariate import SquaredExponentialModel,
   ...: NuggetModel
   ...: from debiased_spatial_whittle.grids.base import RectangularGrid
   ...: from debiased_spatial_whittle.sampling.simulation import SamplerOnRectangularGri
   ...: d
   ...: from debiased_spatial_whittle.inference.periodogram import Periodogram, Expected
   ...: Periodogram
   ...: from debiased_spatial_whittle.inference.likelihood import Estimator, DebiasedWhi
   ...: ttle
   ...: from debiased_spatial_whittle.grids.old import ImgGrid

In [4]: model = SquaredExponentialModel(rho=7, sigma=0.9)

In [5]: shape = (128, 128)
   ...: mask_france = ImgGrid(shape).get_new().astype(bool)
   ...: grid_france = RectangularGrid(shape)
   ...: grid_france.mask = mask_france
   ...: sampler = SamplerOnRectangularGrid(model, grid_france)
FileNotFoundError                         Traceback (most recent call last)
Cell In[5], line 2
      1 shape = (128, 128)
----> 2 mask_france = ImgGrid(shape).get_new().astype(bool)
      3 grid_france = RectangularGrid(shape)
      4 grid_france.mask = mask_france

File ~/mambaforge/envs/temp/lib/python3.9/site-packages/debiased_spatial_whittle/grids/old.py:88, in ImgGrid.__init__(self, shape, img_path)
     86 super().__init__(shape)
     87 self.img_path = img_path
---> 88 img = plt.imread(self.img_path)
     89 img = np.array(img)
     90 img = (img[110:-110, 110:-110, 0] == 0) * 1.0

File ~/mambaforge/envs/temp/lib/python3.9/site-packages/matplotlib/pyplot.py:2597, in imread(fname, format)
   2593 @_copy_docstring_and_deprecators(matplotlib.image.imread)
   2594 def imread(
   2595         fname: str | pathlib.Path | BinaryIO, format: str | None = None
   2596 ) -> np.ndarray:
-> 2597     return matplotlib.image.imread(fname, format)

File ~/mambaforge/envs/temp/lib/python3.9/site-packages/matplotlib/image.py:1544, in imread(fname, format)
   1537 if isinstance(fname, str) and len(parse.urlparse(fname).scheme) > 1:
   1538     # Pillow doesn't handle URLs directly.
   1539     raise ValueError(
   1540         "Please open the URL for reading and pass the "
   1541         "result to Pillow, e.g. with "
   1542         "``np.array(PIL.Image.open(urllib.request.urlopen(url)))``."
   1543         )
-> 1544 with img_open(fname) as image:
   1545     return (_pil_png_to_float_array(image)
   1546             if isinstance(image, PIL.PngImagePlugin.PngImageFile) else
   1547             pil_to_array(image))

File ~/mambaforge/envs/temp/lib/python3.9/site-packages/PIL/Image.py:3513, in open(fp, mode, formats)
   3511 if is_path(fp):
   3512     filename = os.fspath(fp)
-> 3513     fp = builtins.open(filename, "rb")
   3514     exclusive_fp = True
   3515 else:

FileNotFoundError: [Errno 2] No such file or directory: '/home/weiji/mambaforge/envs/temp/lib/python3.9/france.jpg'
```

The path to that file seems to be hardcoded here:

https://github.com/arthurBarthe/debiased-spatial-whittle/blob/e8bd46e01824b44e6c4a6a57b3b19e0cb26865fb/src/debiased_spatial_whittle/grids/old.py#L18

Could you maybe fix the default filepath somehow? Or is it recommended to manually download the france.jpg file and pass the path to that instead?
