### 2025-11-27 #8323 @gkthiruvathukal
@arthurBarthe Version bumps are fine as long as the reviewers are aware of the updates, which they seem to be.

@MarineChap, @weiji14 Please let me know once your review is compete and you are satisfied with the changes being made by @arthurBarthe.
