### 2026-02-19 #8323 @gkthiruvathukal
👋  @arthurBarthe  - we are almost there! Next is just setting up the archive for your new release.

So here is what we have left to do:

- [x]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted in your code repository.

- [x]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly including the case) and the author list (should match the people and the order in which they are listed on your paper exactly).

- [x]  Please respond with the DOI of the archived version here along with the version tag name released.
