### 2026-03-04 #8323 @gkthiruvathukal
@arthurBarthe I think we are almost good to go. There is one minor issue with one of your references:

Can you please take a look and see if you can fix it?

- 10.32614/cran.package.gpgp may be a valid DOI for title: GpGp: Fast Gaussian Process Computation Using Vecc...
