### 2026-03-10 #8323 @arthurBarthe
@gkthiruvathukal Thank you, I have added the missing doi to the corresponding reference in the paper. I will update the zenodo deposit accordingly, however right now there seems to be temporary issues with their website. I will update here when I have done this.
