### 2026-03-24 #8323 @crvernon
@arthurBarthe - I just need you to address the following before I accept for publication:

In your archive:

- [x] There seems to be a version discrepancy between what was declared here in the review thread (1.0.1), what is specified in the Zenodo archive (2.1.3), and what is in the GitHub release (2.1.4).  Which of these versions is correct?  The GitHub release and the Zenodo archive should match what I declare here as the version.

In your paper:

- [x] LINE 92:  The "h" in "Github" should be capitalized.
- [x] LINE 118:  The "g" in "gaussian" should be capitalized.  You can do this in your bib file by placing characters you wish to maintain the formatting of in curly brackets.
- [x] LINE 120:  The "g" in "gaussian" should be capitalized.
- [x] LINE 122:  The "g" in "gaussian" should be capitalized.
- [x] LINE 144:  The  "p" in "python" should be capitalized.

No new release is necessary after making these changes.

Let me know when they have been completed and I'll accept after confirmation.  Thank you.
