### 2026-03-24 #8323 @arthurBarthe
@crvernon Thank you for spotting these issues. The correct version is 2.1.4, following changes made after the initial submission, including some requested by the reviewers. I have updated the version specified in the Zenodo archive to match this.
