### 2026-03-27 #8323 @arthurBarthe
Amazing @crvernon ! Thank you all @gkthiruvathukal @weiji14 @MarineChap  for your contribution in the process, it was my first time publishing with JOSS and I must say I've really found the overall process quite refreshing! I will volunteer as a reviewer myself, and try to advocate publishing research code with JOSS within my network.
